edit: better pending integration
as 9936c8885f37da5f24b42bb33692b577dc4c7c0c is throwing an error if the
table is not rendered at the start and it had to be remebered to add
the `v-if`-tag, this is now a better attempt for integrating the pending
feature.

fore the props given to the datatable component there is now a new prop
called `pending` which enables the table-slot, while hiding (via v-show)
the table. By this attempt the `tbl`-comonent is still beeing rendered.